move component to /Components
